AsyncWorker/requirements.txt
django
djangorestframework
psycopg2-binary
celery
pytesseract 
pillow 
pdf2image
pika
opencv-python
django-celery-results
elasticsearch<9.0.0
sentence-transformers